fix time sync issues by switching to double for all pupil timestamps

+++ plugin/Scripts/Calibration.cs
@@ -71,7 +71,7 @@ namespace PupilLabs
             calibrationData.Clear();
         }
 
-        public void AddCalibrationPointReferencePosition(float[] position, float timestamp)
+        public void AddCalibrationPointReferencePosition(float[] position, double timestamp)
         {
             calibrationData.Add(new Dictionary<string, object>() {
                 { settings.PositionKey, position },
